% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_token.R
\name{dropbox_token}
\alias{dropbox_token}
\title{generate dropbox token
source form https://stackoverflow.com/questions/71393752/get-a-refresh-token-for-dropbox-api-using-rdrop2-and-drop-auth}
\usage{
dropbox_token(
  new_user = FALSE,
  dir = NULL,
  key = "mmhfsybffdom42w",
  secret = "l8zeqqqgm1ne5z0",
  cache = TRUE,
  rdstoken = NA
)
}
\arguments{
\item{new_user}{if TRUE is means generate a new token}

\item{dir}{set directory of token file.}
}
\value{
A token rds file.
}
\description{
generate dropbox token
source form https://stackoverflow.com/questions/71393752/get-a-refresh-token-for-dropbox-api-using-rdrop2-and-drop-auth
}
