% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shiny_app.R
\name{generate_shiny_app}
\alias{generate_shiny_app}
\title{Generate Shiny app}
\usage{
generate_shiny_app(dir = NULL, name = "shiny_app", create_token = T)
}
\arguments{
\item{dir}{set directory of output files.}

\item{name}{set name of Shiny app.}
}
\value{
A Shiny app file,logo pic and report Rmd files.
}
\description{
This function launches a simple Shiny app with a text input and output.
}
