% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedrive_token.R
\name{onedrive_token}
\alias{onedrive_token}
\title{generate OneDrive token}
\usage{
onedrive_token(dir = NULL, new_user = F)
}
\arguments{
\item{dir}{set directory of token file.}
}
\value{
A token rds file.
}
\description{
OneDrive
}
