% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebTrials-package.R
\docType{package}
\name{WebTrials-package}
\alias{WebTrials}
\alias{WebTrials-package}
\title{WebTrials: 'WebTrials' Template Shiny Web For Clinical Trials}
\description{
Create Template Shiny Web For Clinical Trials
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maemod-moru.github.io/WebTrials/}
}

}
\keyword{internal}
